+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 * added support for overriding the snapshot path.
 * added an `--accept` command to `cargo insta test`
 * correctly handle nested macros that might contain inline snapshots.
+* use thread name as snapshot name for inline snapshots.